+++ src/components/views/GetViewInputsView.jsx
@@ -162,7 +162,7 @@ export function GetViewInputsView({ currentContext = null, instance: viewInstanc
   return (
     <DataList
       currentContext={currentContext}
-      headerActions={['openAll', 'copy', 'reload', 'refresh']}
+      headerActions={['openAll', 'reload', 'refresh']}
       isRefreshing={isRefreshing}
       itemActions={['copy', 'openAll', 'viewsExplorer']}
       itemLabel='dataset'